feat: Create PersonHeader and Refactor
